--- app/src/main/java/com/bintina/goouttolunchmvvm/login/view/UserActivity.kt
@@ -1,4 +1,0 @@
-package com.bintina.goouttolunchmvvm.login.view
-
-class UserActivity {
-}
