--- app/src/main/java/com/bintina/goouttolunchmvvm/login/viewmodel/LoginViewModelFactory.kt
@@ -1,15 +1,0 @@
-package com.bintina.goouttolunchmvvm.login.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.viewmodel.CreationExtras
-
-class LoginViewModelFactory: ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>, extras: CreationExtras): T {
-        if(modelClass.isAssignableFrom(LogInViewModel::class.java)){
-            return LogInViewModel() as T
-        }
-        throw IllegalArgumentException("ViewModel not found")
-    }
-
-}
